docs: Add Javadoc, Rename Duke
Add Javadoc to all non-private classes and methods.

Rename all instances of Duke to Prts, this has no outward-facing
effects on the user, and no other effects on development.
