+++ build.gradle
@@ -28,11 +28,11 @@ test {
 }
 
 application {
-    mainClass.set("duke.Duke")
+    mainClass.set("prts.Prts")
 }
 
 shadowJar {
-    archiveBaseName = "duke"
+    archiveBaseName = "prts"
     archiveClassifier = null
     dependsOn("distZip", "distTar")
 }
