--- src/main/java/duke/CreateNewSaveException.java
@@ -1,7 +1,0 @@
-package duke;
-
-public class CreateNewSaveException extends DukeException {
-    public CreateNewSaveException() {
-        super("File failed to load, created new save file.");
-    }
-}
