--- src/main/java/duke/OutOfRangeException.java
@@ -1,7 +1,0 @@
-package duke;
-
-public class OutOfRangeException extends DukeException {
-    public OutOfRangeException() {
-        super("Please provide a number within range.");
-    }
-}
