--- src/main/java/duke/SaveToFileException.java
@@ -1,7 +1,0 @@
-package duke;
-
-public class SaveToFileException extends DukeException {
-    public SaveToFileException() {
-        super("Failed to save to save file.");
-    }
-}
