--- src/main/java/duke/command/ExitCommand.java
@@ -1,17 +1,0 @@
-package duke.command;
-
-import duke.Storage;
-import duke.TaskList;
-import duke.Ui;
-
-public class ExitCommand extends Command {
-
-    @Override
-    public void execute(TaskList tasks, Ui ui, Storage storage) {}
-
-    @Override
-    public boolean isExit() {
-        return true;
-    }
-
-}
