--- src/main/java/duke/command/MessageCommand.java
@@ -1,16 +1,0 @@
-package duke.command;
-
-import duke.Storage;
-import duke.TaskList;
-import duke.Ui;
-
-public class MessageCommand extends Command {
-    private String message;
-    public MessageCommand(String message) {
-        this.message = message;
-    }
-    @Override
-    public void execute(TaskList tasks, Ui ui, Storage storage) {
-        ui.displayMessage(message);
-    }
-}
