--- src/main/java/duke/task/AlreadyMarkedException.java
@@ -1,9 +1,0 @@
-package duke.task;
-
-import duke.DukeException;
-
-public class AlreadyMarkedException extends DukeException {
-    public AlreadyMarkedException() {
-        super("It's already done.");
-    }
-}
