--- src/main/java/duke/task/AlreadyUnmarkedException.java
@@ -1,9 +1,0 @@
-package duke.task;
-
-import duke.DukeException;
-
-public class AlreadyUnmarkedException extends DukeException {
-    public AlreadyUnmarkedException() {
-        super("This hasn't been done yet...");
-    }
-}
