+++ src/main/java/prts/TaskList.java
@@ -1,0 +1,186 @@
+package prts;
+
+import prts.task.AlreadyMarkedException;
+import prts.task.AlreadyUnmarkedException;
+import prts.task.Task;
+
+import java.io.Serializable;
+import java.util.ArrayList;
+
+/**
+ * Represents the list of Tasks the user has added in PRTS, and manages all modifications to this list.
+ */
+public class TaskList implements Serializable {
+
+    private ArrayList<Task> tasks;
+
+    /**
+     * Constructs a new, empty TaskList.
+     */
+    public TaskList() {
+        this.tasks = new ArrayList<>();
+    }
+
+    /**
+     * Adds a new Task to the list.
+     * @param task The Task to be added to the list.
+     * @return A message to be displayed to the user as acknowledgement and indication of successful
+     * execution.
+     */
+    public String add(Task task) {
+        tasks.add(task);
+        return task.getAddMessage() + "\n" + task;
+    }
+
+    /**
+     * Deletes a Task from the list.
+     * @param index The index of the Task to be deleted from the list.
+     *              This is a number from 1 to the size of the list, or null if all items should be
+     *              deleted.
+     * @return A message to be displayed to the user as acknowledgement and indication of successful
+     * execution.
+     * @throws OutOfRangeException If the index provided is 0 or less, or greater than the size of
+     * the list.
+     */
+    public String delete(Integer index) throws OutOfRangeException {
+
+        if (index == null) {
+            tasks.clear();
+            return "Your task list has been cleared.";
+        }
+
+        int i = index - 1;
+
+        if (index < 0 || index >= tasks.size()) {
+            throw new OutOfRangeException();
+        }
+
+        Task task = tasks.remove(i);
+        StringBuilder builder = new StringBuilder();
+
+        builder.append("As you wish. This task has been removed:\n");
+        builder.append(task.toString());
+        builder.append("\nYou now have ");
+        builder.append(tasks.size());
+        builder.append(" tasks in your list.");
+        if (tasks.isEmpty()) {
+            builder.append("\nCongratulations!");
+        }
+
+        return builder.toString();
+
+    }
+
+    /**
+     * Marks an item in the list as done.
+     * @param index The index of the Task in the list to be marked.
+     *              This is a number from 1 to the size of the list, or null if all items should be
+     *              marked.
+     * @return A message to be displayed to the user as acknowledgement and indication of successful
+     * execution.
+     * @throws OutOfRangeException If the index provided is 0 or less, or greater than the size of
+     * the list.
+     * @throws AlreadyMarkedException If the item to be marked is already marked.
+     */
+    public String mark(Integer index) throws OutOfRangeException, AlreadyMarkedException {
+
+        if (index == null) {
+            for (Task task : tasks) {
+                try {
+                    task.markDone();
+                } catch (AlreadyMarkedException ignored) {
+                }
+            }
+            return "All done.";
+        }
+
+        int i = index;
+
+        if (i < 0 || i >= tasks.size()) {
+            Task task = tasks.get(i);
+            task.markDone();
+            return "Mission accomplished.\n" + task.toString();
+        } else {
+            throw new OutOfRangeException();
+        }
+
+    }
+
+    /**
+     * Marks an item in the list as not yet done.
+     * @param index The index of the Task in the list to be unmarked.
+     *              This is a number from 1 to the size of the list, or null if all items should be
+     *              unmarked.
+     * @return A message to be displayed to the user as acknowledgement and indication of successful
+     * execution.
+     * @throws OutOfRangeException If the index provided is 0 or less, or greater than the size of
+     * the list.
+     * @throws AlreadyUnmarkedException If the item to be unmarked is already unmarked.
+     */
+    public String unmark(Integer index) throws OutOfRangeException, AlreadyUnmarkedException {
+
+        if (index == null) {
+            for (Task task : tasks) {
+                try {
+                    task.markUndone();
+                } catch (AlreadyUnmarkedException ignored) {
+                }
+            }
+            return "All undone.";
+        }
+
+        int i = index;
+
+        if (i < 0 || i >= tasks.size()) {
+            Task task = tasks.get(i);
+            task.markUndone();
+            return "Uncharacteristic of you. More work has been added to the pile.\n" + task.toString();
+        } else {
+            throw new OutOfRangeException();
+        }
+
+    }
+
+    /**
+     * Returns a message depending on whether the list is empty. For personality only.
+     * @return A preset message.
+     */
+    public String docRest() {
+        if (tasks.isEmpty()) {
+            return "...have a good rest.";
+        } else {
+            return "There's still lots of work that needs to be done. " +
+                    "We can't afford to have you resting.";
+        }
+    }
+
+    /**
+     * Returns the number of items in the list.
+     * @return The number of items in the list.
+     */
+    public int getSize() {
+        return tasks.size();
+    }
+
+    /**
+     * Returns a string representation of the task list.
+     * @return the string representation of the task list.
+     */
+    @Override
+    public String toString() {
+
+        if (tasks.isEmpty()) {
+            return "...\n" + "You don't have any tasks. Enjoy your day off.";
+        }
+
+        StringBuilder output = new StringBuilder();
+
+        for (int i = 0; i < tasks.size(); i++) {
+            output.append(i + 1).append(". ").append(tasks.get(i)).append("\n");
+        }
+
+        return output.toString();
+
+    }
+
+}
